consolidate words to one plugin
